Refactored Tests to Tests folder
