--- Assets/Editor/GameTests.cs
@@ -1,13 +1,0 @@
-﻿using NUnit.Framework;
-
-namespace Assets.Editor
-{
-    public class MainMenuTestsa
-    {
-        public class Exit_Button_Pressed_Method
-        {
-         
-        
-        }
-    }
-}
