--- Assets/Editor/GameTests.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: d25027ef5112a544c85fd6acd73a51c8
-timeCreated: 1517149980
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
