+++ Assets/Tests.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 464e3ff2f2e7b70458c9ab0fbd426230
+folderAsset: yes
+timeCreated: 1517278934
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
